feat: harden Study Mode quiz experience
